requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
asyncio-mqtt>=0.16.0
aiofiles>=23.2.1
httpx>=0.25.0
uvloop>=0.19.0

# AI and ML Framework
dspy>=3.0.0
openai>=1.12.0
google-generativeai>=0.3.0
anthropic>=0.8.0

# MCP (Model Context Protocol)
mcp>=1.0.0
mcp-server-memory>=1.0.0
mcp-server-filesystem>=1.0.0
mcp-server-git>=1.0.0

# Vector Database and Memory
chromadb>=0.4.18
redis>=5.0.1
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Web Framework and UI
streamlit>=1.29.0
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=12.0

# Data Processing
pandas>=2.1.4
numpy>=1.24.3
python-dateutil>=2.8.2
pytz>=2023.3

# Security and Encryption
cryptography>=41.0.8
keyring>=24.3.0
python-jose>=3.3.0
passlib>=1.7.4

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0
sentry-sdk>=1.38.0

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Utilities
click>=8.1.7
rich>=13.7.0
tqdm>=4.66.1
schedule>=1.2.0
watchdog>=3.0.0

# Optional: Voice and Audio (uncomment if needed)
# speech-recognition>=3.10.0
# pyttsx3>=2.90
# pyaudio>=0.2.11

# Optional: Additional MCP Servers (uncomment as needed)
# mcp-server-gmail>=1.0.0
# mcp-server-notion>=1.0.0
# mcp-server-github>=1.0.0
# mcp-server-slack>=1.0.0
# mcp-server-discord>=1.0.0
